+++ src/routes.ts
@@ -24,7 +24,6 @@ import { AccountPendingInvitationsPage } from "./pages/account_pending_invitatio
 import { StudentCourseInvitationPage } from "./pages/student_course_invitation_page";
 import { StudentCourseInvitationsPage } from "./pages/student_course_invitations_page";
 import { StudentCourseLandingPage } from "./pages/student_course_landing_page";
-import { StudentCoursesPage } from "./pages/student_courses_page";
 import { TeachingOperationsPage } from "./pages/teaching_operations_page";
 import { InstructorAccountsPage } from "./pages/instructor_accounts_page";
 import { SupportRosterPage } from "./pages/support_roster_page";
@@ -45,7 +44,6 @@ const routeComponents: Readonly<Record<RouteId, Component>> = {
   courses: CourseListPage,
   signIn: SignInPage,
   pendingCourseInvitations: AccountPendingInvitationsPage,
-  studentCourses: StudentCoursesPage,
   studentCourseInvitations: StudentCourseInvitationsPage,
   studentCourseInvitation: StudentCourseInvitationPage,
   studentCourseLanding: StudentCourseLandingPage,